fix realtime stats (sniper alive time counting): division by zero. Now it should work perfectly fix typo on fairness.cfg (remeber: changed name of dvar from scr_enable_spawn_protection to scr_spawn_protection_enable) fix: level.scr_player_maxhealth doesn't exist, it's level.maxhealth setting to show the crosshair in hardcore setting to allow grenade cooking (needed for promod emulation too):   added new weapon: frag_grenade_nocook_mp for that behavior, when cooking is disabled, this type of grenade is given instead   TODO: default classes don't respect disabled cooking (they always use frag_grenade_mp). It will be fixed after changing tablelookups to .cfg files
re-enabled some ow logic files (hoping that it would fix scr_oldschool not moving):
  _antibunnyhopping.gsc
  _fitnesscs.gsc
  _weaponweightmodifier
  _extendedobituaries.gsc

moved text from You killed X a little bellow because _extendedobituaries.gsc text would be on the top of it. Moved text from objective capture a little below too, so it wouldn't interfere with the new position of "you killed"

if in a game, when in menu createserver, Fast Restart option is disabled if the gametype to load is different from the current one. This is to avoid script error on _spawnlogic